+++ tests/test_parsers.py
@@ -4,7 +4,7 @@ from __future__ import absolute_import
 from __future__ import print_function
 from __future__ import unicode_literals
 import unittest
-from asciimatics.parsers import AsciimaticsParser, AnsiTerminalParser, ControlCodeParser
+from asciimatics.parsers import AsciimaticsParser, AnsiTerminalParser, ControlCodeParser, Parser
 import asciimatics.constants as constants
 
 
@@ -16,13 +16,12 @@ class TestParsers(unittest.TestCase):
         """
         parser = ControlCodeParser()
         parser.reset("\0\b\ra[", None)
-        self.assertEquals(parser.normalize(), "\0\b\ra[")
         tokens = parser.parse()
-        self.assertEquals(next(tokens), ("^@", (None, None, None), 0))
-        self.assertEquals(next(tokens), ("^H", (None, None, None), 1))
-        self.assertEquals(next(tokens), ("^M", (None, None, None), 2))
-        self.assertEquals(next(tokens), ("a", (None, None, None), 3))
-        self.assertEquals(next(tokens), ("[", (None, None, None), 4))
+        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "^@"))
+        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "^H"))
+        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "^M"))
+        self.assertEquals(next(tokens), (3, Parser.DISPLAY_TEXT, "a"))
+        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "["))
 
     def test_asciimatics_parser(self):
         """
@@ -30,14 +29,16 @@ class TestParsers(unittest.TestCase):
         """
         parser = AsciimaticsParser()
         parser.reset("a${1}b${2,1}c${3,2,4}de${7}", None)
-        self.assertEquals(parser.normalize(), "a${1}b${2,1}c${3,2,4}de${7}")
         tokens = parser.parse()
-        self.assertEquals(next(tokens), ("a", (None, None, None), 0))
-        self.assertEquals(next(tokens), ("b", (1, 0, None), 1))
-        self.assertEquals(next(tokens), ("c", (2, 1, None), 6))
-        self.assertEquals(next(tokens), ("d", (3, 2, 4), 13))
-        self.assertEquals(next(tokens), ("e", (3, 2, 4), 22))
-        self.assertEquals(next(tokens), (None, (7, 0, None), 23))
+        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEquals(next(tokens), (1, Parser.CHANGE_COLOURS, (1, 0, None)))
+        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEquals(next(tokens), (6, Parser.CHANGE_COLOURS, (2, 1, None)))
+        self.assertEquals(next(tokens), (6, Parser.DISPLAY_TEXT, "c"))
+        self.assertEquals(next(tokens), (13, Parser.CHANGE_COLOURS, (3, 2, 4)))
+        self.assertEquals(next(tokens), (13, Parser.DISPLAY_TEXT, "d"))
+        self.assertEquals(next(tokens), (22, Parser.DISPLAY_TEXT, "e"))
+        self.assertEquals(next(tokens), (23, Parser.CHANGE_COLOURS, (7, 0, None)))
         with self.assertRaises(StopIteration):
             next(tokens)
 
@@ -50,31 +51,37 @@ class TestParsers(unittest.TestCase):
         tokens = parser.parse()
 
         # Normal text
-        self.assertEquals(next(tokens), ("a", (None, None, None), 0))
+        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
 
         # Unknown escape code
-        self.assertEquals(next(tokens), ("b", (None, None, None), 1))
+        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
 
         # Reset
-        self.assertEquals(next(tokens), ("c", (7, constants.A_NORMAL, 0), 7))
+        self.assertEquals(next(tokens), (7, Parser.CHANGE_COLOURS, (7, constants.A_NORMAL, 0)))
+        self.assertEquals(next(tokens), (7, Parser.DISPLAY_TEXT, "c"))
 
         # Bold
-        self.assertEquals(next(tokens), ("d", (7, constants.A_BOLD, 0), 12))
+        self.assertEquals(next(tokens), (12, Parser.CHANGE_COLOURS, (7, constants.A_BOLD, 0)))
+        self.assertEquals(next(tokens), (12, Parser.DISPLAY_TEXT, "d"))
 
         # Normal
-        self.assertEquals(next(tokens), ("e", (7, constants.A_NORMAL, 0), 17))
+        self.assertEquals(next(tokens), (17, Parser.CHANGE_COLOURS, (7, constants.A_NORMAL, 0)))
+        self.assertEquals(next(tokens), (17, Parser.DISPLAY_TEXT, "e"))
 
         # Inverse
-        self.assertEquals(next(tokens), ("f", (7, constants.A_REVERSE, 0), 22))
+        self.assertEquals(next(tokens), (22, Parser.CHANGE_COLOURS, (7, constants.A_REVERSE, 0)))
+        self.assertEquals(next(tokens), (22, Parser.DISPLAY_TEXT, "f"))
 
         # Unset inverse
-        self.assertEquals(next(tokens), ("g", (7, constants.A_NORMAL, 0), 27))
+        self.assertEquals(next(tokens), (27, Parser.CHANGE_COLOURS, (7, constants.A_NORMAL, 0)))
+        self.assertEquals(next(tokens), (27, Parser.DISPLAY_TEXT, "g"))
 
         # Standard colours, using multiple parameters
-        self.assertEquals(next(tokens), ("h", (constants.COLOUR_RED, constants.A_NORMAL, constants.COLOUR_GREEN), 33))
+        self.assertEquals(next(tokens), (33, Parser.CHANGE_COLOURS, (constants.COLOUR_RED, constants.A_NORMAL, constants.COLOUR_GREEN)))
+        self.assertEquals(next(tokens), (33, Parser.DISPLAY_TEXT, "h"))
 
         # Final escape sequence with no visible text is returned with no text.
-        self.assertEquals(next(tokens), (None, (constants.COLOUR_WHITE, constants.A_NORMAL, constants.COLOUR_BLACK), 42))
+        self.assertEquals(next(tokens), (42, Parser.CHANGE_COLOURS, (constants.COLOUR_WHITE, constants.A_NORMAL, constants.COLOUR_BLACK)))
 
         with self.assertRaises(StopIteration):
             next(tokens)
@@ -84,46 +91,73 @@ class TestParsers(unittest.TestCase):
         Check AnsiTerminalParser colour palettes work as expected.
         """
         parser = AnsiTerminalParser()
-        parser.reset("\x1B[38;1ma\x1B[38;5;17mb\x1B[48;2;1;2;3mc\x1B[48;5;54md\x1B[999me", None)
+        parser.reset(
+            "\x1B[38;1ma\x1B[38;5;17mb\x1B[48;2;1;2;3mc\x1B[48;5;54md\x1B[999me\x1B[93m\x1B[104m", None)
         tokens = parser.parse()
 
         # Bad colour scheme - ignore
-        self.assertEquals(next(tokens), ("a", (None, None, None), 0))
+        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
 
         # Standard colour palette
-        self.assertEquals(next(tokens), ("b", (17, None, None), 8))
+        self.assertEquals(next(tokens), (8, Parser.CHANGE_COLOURS, (17, None, None)))
+        self.assertEquals(next(tokens), (8, Parser.DISPLAY_TEXT, "b"))
 
         # RGB colour scheme - ignore
-        self.assertEquals(next(tokens), ("c", (17, None, None), 19))
+        self.assertEquals(next(tokens), (19, Parser.DISPLAY_TEXT, "c"))
 
         # Standard colour palette
-        self.assertEquals(next(tokens), ("d", (17, None, 54), 33))
+        self.assertEquals(next(tokens), (33, Parser.CHANGE_COLOURS, (17, None, 54)))
+        self.assertEquals(next(tokens), (33, Parser.DISPLAY_TEXT, "d"))
 
         # Unknown parameter
-        self.assertEquals(next(tokens), ("e", (17, None, 54), 44))
+        self.assertEquals(next(tokens), (44, Parser.DISPLAY_TEXT, "e"))
+
+        # Intense colour palette
+        self.assertEquals(next(tokens), (51, Parser.CHANGE_COLOURS, (11, None, 54)))
+        self.assertEquals(next(tokens), (51, Parser.CHANGE_COLOURS, (11, None, 12)))
 
     def test_ansi_terminal_parser_cursor(self):
         """
         Check AnsiTerminalParser cursor movement work as expected.
         """
         parser = AnsiTerminalParser()
-        parser.reset("aa\x08b\rc\x1B[Cdd\x1B[De\r", None)
+        parser.reset("aa\x08b\rc\x1B[Cdd\x1B[De\x1B[A\x1B[B\x1B[1;2H\x1B[?25h\x1B[?25l\r", None)
         tokens = parser.parse()
 
-        # Carriage return and overwrite
-        self.assertEquals(next(tokens), ("c", (None, None, None), 4))
+        # Normal text...
+        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "a"))
 
         # Backspace and overwrite.
-        self.assertEquals(next(tokens), ("b", (None, None, None), 2))
+        self.assertEquals(next(tokens), (2, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "b"))
+
+        # Carriage return and overwrite
+        self.assertEquals(next(tokens), (4, Parser.MOVE_ABSOLUTE, (0, None)))
+        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "c"))
 
         # Move cursor forwards and append.
-        self.assertEquals(next(tokens), ("d", (None, None, None), 6))
+        self.assertEquals(next(tokens), (6, Parser.MOVE_RELATIVE, (1, 0)))
+        self.assertEquals(next(tokens), (6, Parser.DISPLAY_TEXT, "d"))
+        self.assertEquals(next(tokens), (10, Parser.DISPLAY_TEXT, "d"))
 
         # Move cursor backwards and overwrite.
-        self.assertEquals(next(tokens), ("e", (None, None, None), 11))
+        self.assertEquals(next(tokens), (11, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEquals(next(tokens), (11, Parser.DISPLAY_TEXT, "e"))
 
-        # Normalize returns correct linear form - complete with accurate cursor location.
-        self.assertEqual(parser.normalize(), "cbde\x1B[4D")
+        # Move cursor up and down.
+        self.assertEquals(next(tokens), (15, Parser.MOVE_RELATIVE, (0, -1)))
+        self.assertEquals(next(tokens), (15, Parser.MOVE_RELATIVE, (0, 1)))
+
+        # Move cursor to location
+        self.assertEquals(next(tokens), (15, Parser.MOVE_ABSOLUTE, (1, 0)))
+
+        # Show/hide cursor
+        self.assertEquals(next(tokens), (15, Parser.SHOW_CURSOR, True))
+        self.assertEquals(next(tokens), (15, Parser.SHOW_CURSOR, False))
+
+        # Trailing Carriage return
+        self.assertEquals(next(tokens), (15, Parser.MOVE_ABSOLUTE, (0, None)))
 
     def test_ansi_terminal_parser_delete(self):
         """
@@ -134,57 +168,56 @@ class TestParsers(unittest.TestCase):
         # Delete to end of line
         parser.reset("abcde\x08\x08\x08\x1B[K", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), ("a", (None, None, None), 0))
-        self.assertEquals(next(tokens), ("b", (None, None, None), 1))
-        self.assertEquals(next(tokens), (None, (None, None, None), 5))
+        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
+        self.assertEquals(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
+        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
+        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEquals(next(tokens), (5, Parser.DELETE_LINE, 0))
         with self.assertRaises(StopIteration):
             next(tokens)
 
         # Delete to start of line
-        parser.reset("abcde\x08\x08\x08\x1B[1K", None)
+        parser.reset("abcde\x1B[1K", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), (" ", (None, None, None), 8))
-        self.assertEquals(next(tokens), (" ", (None, None, None), 8))
-        self.assertEquals(next(tokens), ("c", (None, None, None), 2))
-        self.assertEquals(next(tokens), ("d", (None, None, None), 3))
-        self.assertEquals(next(tokens), ("e", (None, None, None), 4))
-        self.assertEquals(next(tokens), (None, (None, None, None), 5))
+        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
+        self.assertEquals(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
+        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
+        self.assertEquals(next(tokens), (5, Parser.DELETE_LINE, 1))
         with self.assertRaises(StopIteration):
             next(tokens)
 
         # Delete line
-        parser.reset("abcde\x08\x08\x08\x1B[2K", None)
+        parser.reset("abcde\x1B[2K", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), (" ", (None, None, None), 8))
-        self.assertEquals(next(tokens), (" ", (None, None, None), 8))
-        self.assertEquals(next(tokens), (None, (None, None, None), 5))
+        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
+        self.assertEquals(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
+        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
+        self.assertEquals(next(tokens), (5, Parser.DELETE_LINE, 2))
         with self.assertRaises(StopIteration):
             next(tokens)
 
         # Delete char
         parser.reset("abcde\x08\x08\x08\x1B[P", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), ("a", (None, None, None), 0))
-        self.assertEquals(next(tokens), ("b", (None, None, None), 1))
-        self.assertEquals(next(tokens), ("d", (None, None, None), 3))
-        self.assertEquals(next(tokens), ("e", (None, None, None), 4))
-        self.assertEquals(next(tokens), (None, (None, None, None), 5))
+        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
+        self.assertEquals(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
+        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
+        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEquals(next(tokens), (5, Parser.DELETE_CHARS, 1))
         with self.assertRaises(StopIteration):
             next(tokens)
-
-    def test_ansi_terminal_parser_normalization(self):
-        """
-        Check AnsiTerminalParser normalization works as expected.
-        """
-        parser = AnsiTerminalParser()
-
-        # SGR0 sets black and white normal text.
-        parser.reset("\x1B[ma", None)
-        self.assertEquals(parser.normalize(), "\x1B[38;5;7;2;48;5;0ma")
-
-        # SGR1 sets bold and SGR7 reverse video.
-        parser.reset("\x1B[1ma\x1B[7mb", None)
-        self.assertEquals(parser.normalize(), "\x1B[1ma\x1B[7mb")
 
     def test_ansi_terminal_parser_errors(self):
         """
@@ -195,8 +228,37 @@ class TestParsers(unittest.TestCase):
         tokens = parser.parse()
 
         # Ignore unknown escape and next letter
-        self.assertEquals(next(tokens), ("a", (None, None, None), 0))
-        self.assertEquals(next(tokens), ("b", (None, None, None), 1))
+        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
 
         # Ignore unknown control char
-        self.assertEquals(next(tokens), ("c", (None, None, None), 4))
+        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "c"))
+
+    def test_ansi_terminal_parser_tab(self):
+        """
+        Check AnsiTerminalParser handles tabs.
+        """
+        parser = AnsiTerminalParser()
+        parser.reset("\x09", None)
+        tokens = parser.parse()
+        self.assertEquals(next(tokens), (0, Parser.NEXT_TAB, None))
+
+    def test_ansi_terminal_parser_clear(self):
+        """
+        Check AnsiTerminalParser clears screen.
+        """
+        parser = AnsiTerminalParser()
+        parser.reset("\x1B[2J", None)
+        tokens = parser.parse()
+        self.assertEquals(next(tokens), (0, Parser.CLEAR_SCREEN, None))
+
+    def test_ansi_terminal_parser_os_cmd(self):
+        """
+        Check AnsiTerminalParser removes OS commands.
+        """
+        parser = AnsiTerminalParser()
+        parser.reset("a\x1B]do something;stuff:to^ignore\x07b", None)
+        tokens = parser.parse()
+        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+