+++ CHANGES.rst
@@ -9,12 +9,15 @@ LATEST
 - Added readonly logic for Text and TextBox.
 - Added ability to enable/disable widgets by column in layouts.
 - Added left/right/up/down navigation to nearest widget.
+- Added ability to scroll screen/canvas by variable number of lines.
+- Created terminal demo
 - Fixed exception on reinstating NoneType signal handler.
 - Fixed float/int issue with recent builds of pywin32.
 - Fixed issue where setting options changed the selected value (even if it was still present).
 - Fixed erroneous trigger of on_load for all Frames at start of day.
 - Fixed bug where Frames passed on events that they already handled.
 - Fixed bug: Restore current theme on screen resize.
+- Fixed bug in scrolling the screen up.
 
 1.11.0
 ------